backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.0
python-dotenv==1.0.1
PyMySQL==1.1.0
cryptography==43.0.3
Flask-Cors==4.0.0
pytest==8.3.3
pytest-flask==1.3.0